misc/perf-cron-job/requirements.txt
certifi==2018.4.16
cryptography==2.3
pymongo==3.7.1
PyMySQL==0.9.2
python-dateutil==2.7.3
requests==2.19.1
pyyaml==3.13